change scripts to use conv1d
I have removed the width parameter and use now tf.nn.conv1d. Note
however that there is not tf.nn.conv1d_transpose so I keep on
using tf.nn.conv2d_transpose with a 1 in the third dimension.
The code seems to work in a similar way as the conv2d version.